+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     * you're are more free in your scripts
     * meanwhile, in case of error, you get a less precise info and in english, but you still have the error line
     * all samples work fine now
+    * adapted syntax manual
 
 ## 4.2.14